GEM-ML/requirements.txt
captum==0.6.0
dill==0.3.6
folium==0.14.0
ipywidgets==8.0.6
natsort==8.3.1
eo-learn-core==1.4.2
eo-learn-coregistration==1.4.2
eo-learn-features==1.4.2
eo-learn-geometry==1.4.2
eo-learn-io==1.4.2
eo-learn-mask==1.4.2
eo-learn-ml-tools==1.4.2
eo-learn-visualization==1.4.2
notebook==6.5.4
protobuf<4,>=3.8.0
scikit-learn==1.2.2
scipy<1.9.2,>=1.8
segmentation-models-pytorch==0.3.2
sentinelhub==3.9.1
tdigest==0.5.2.2
tensorboard==2.13.0
tensorboardX==2.6
urllib3<1.27,>=1.25.4